% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{cbind_pops}
\alias{cbind_pops}
\title{cbind_pops}
\usage{
cbind_pops(pop1, pop2)
}
\arguments{
\item{pop1}{first population list, must have named matrices Ni and nit as list elements.}

\item{pop2}{second population list, must have named matrices Ni and nit with same number of rows (sampling sites) as pop1.}
}
\value{
Single population list with sites from pop1 and pop2.
}
\description{
cbind two population lists (as generated from either gen_Nmix_closed or gen_Nmix_open)
}
