% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_reduced_count_models.R
\name{drbinom_normalize}
\alias{drbinom_normalize}
\title{Internal function, should not be called directly. Used to calculate normalization constant for drbinom()}
\usage{
drbinom_normalize(x, size, prob, red, FUN = round)
}
\description{
Internal function, should not be called directly. Used to calculate normalization constant for drbinom()
}
