% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_reduced_count_models.R
\name{drpois_normalize}
\alias{drpois_normalize}
\title{Internal function, should not be called directly. Used to calculate normalization constant for drpois()}
\usage{
drpois_normalize(x, lambda, red, FUN = round)
}
\description{
Internal function, should not be called directly. Used to calculate normalization constant for drpois()
}
