% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{drpoisAPA}
\alias{drpoisAPA}
\title{Reduced poisson probability distribution function \eqn{rPoisson(x;\lambda,r)}, takes reduced quantiles (use drpois2 for full quantiles).}
\usage{
drpoisAPA(x, lambda, red, precBits = 128, log = FALSE)
}
\arguments{
\item{x}{Reduced count quantile (alternatively input reduction(x,r) if x is a full count quantile).}

\item{lambda}{Mean of the full count poisson distribution.}

\item{red}{The factor r by which x was reduced.}

\item{precBits}{Number of bits of precision for arbitrary precision arithmetic.}
}
\value{
The probability of observing quantile \eqn{x}
}
\description{
Reduced poisson probability distribution function \eqn{rPoisson(x;\lambda,r)}, takes reduced quantiles (use drpois2 for full quantiles).
}
\examples{
# probability of observing 105 from pois(lam=90)
x <- drpoisAPA(x = 105, lambda = 90, precBits=64)
}
