% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_reduced_count_models.R
\name{gen_Nmix_closed}
\alias{gen_Nmix_closed}
\title{Generate a population/observation pair with the structure of a closed N-mixture model}
\usage{
gen_Nmix_closed(num_sites, num_times, lambda, pdet)
}
\arguments{
\item{num_sites}{The number of observation sites.}

\item{num_times}{The number of sampling occasions.}

\item{lambda}{The population rate parameter ($N_i \sim Poisson(\lambda)$)}

\item{pdet}{The probability of detection $p$ ($n_{it} \sim Binomial(N_i,p)$)}
}
\value{
A list object with two named matrices. Ni contains the total population per site
        (each row represents a site, each column a sampling occasion). nit contains the observed
        counts (rows=sites, columns=sampling occasions).
}
\description{
Generate a population/observation pair with the structure of a closed N-mixture model
}
\examples{
pop <- gen_Nmix_closed(num_sites = 5, num_times = 10, lambda = 50, pdetection = 0.4)
pop
}
