% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions.R
\name{LR}
\alias{LR}
\title{LR}
\usage{
LR(fullCounts, r = 1)
}
\arguments{
\item{fullCounts}{An R by T matrix of full counts}

\item{r}{The reduction factor to reduce the full counts matrix by}
}
\description{
Calculate maximum loss of precision from considering reduced counts over full counts.
}
\examples{
nit = matrix(c(114,117,122,131,112,100,114,138,113,114,
               116, 96,105,101,109,107,101,111,102, 98,
                99, 90, 92, 93,103, 82,100, 78,100, 81,
                92,130,112,110,114,114,113, 99,109,108,
                99,102,110,105,106,103,101,104, 92,115),
             nrow=5, ncol=10, byrow=T)
LR(nit, r=10)
}
