% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_reduced_count_models.R
\name{mcd}
\alias{mcd}
\title{Find the Most Common Divisor of a data set}
\usage{
mcd(nit, howmany = 10)
}
\arguments{
\item{nit}{Full counts}
}
\description{
Find the Most Common Divisor of a data set
}
\examples{
mcd(c(10,20,30,35,12,18), howmany = 5)

}
