% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_reduced_count_models.R
\name{plot_2d_red_like_closed}
\alias{plot_2d_red_like_closed}
\title{Plot 2D likelihood given a range for pdet and for lambda.}
\usage{
plot_2d_red_like_closed(nit, startPdet, endPdet, stepsizePdet, startLambda,
  endLambda, stepsizeLambda, red, K)
}
\arguments{
\item{nit}{R by T matrix of reduced counts with R sites/rows and T sampling occassions/columns.}

\item{startPdet}{Starting value for pdet.}

\item{endPdet}{Ending value for pdet.}

\item{stepsizePdet}{Spacing between values of pdet}

\item{startLambda}{Starting value for lambda.}

\item{endLambda}{Ending value for lambda.}

\item{stepsizeLambda}{Spacing between values of lambda}

\item{red}{Reduction factor.}

\item{K}{Upper bound on summations (reduced counts upper bound).}
}
\value{
Returns a matrix of likelihoods where rows represent pdet, and columns represent lambda.
}
\description{
Plot 2D likelihood given a range for pdet and for lambda.
}
\examples{
Y <- gen_Nmix_closed(5,5,250,0.5)
plot_2d_red_like_closed(reduction(Y$nit,10), 0.1, 1.0, 0.25, 100, 400, 100, 10, reduction(400,10))
}
