% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_reduced_count_models.R
\name{plot_red_like_closed_lambda}
\alias{plot_red_like_closed_lambda}
\title{Plot likelihood given a pdet and range for lambda.}
\usage{
plot_red_like_closed_lambda(nit, startLambda, endLambda, stepsize, pdet,
  red, K)
}
\arguments{
\item{nit}{R by T matrix of reduced counts with R sites/rows and T sampling occassions/columns.}

\item{startLambda}{Starting value for lambda.}

\item{endLambda}{Ending value for lambda.}

\item{stepsize}{Spacing between values of lambda}

\item{pdet}{Probability of detection (pdet = 0.5 means 50\% chance of detection)}

\item{red}{Reduction factor.}

\item{K}{Upper bound on summations (reduced counts upper bound).}
}
\description{
Plot likelihood given a pdet and range for lambda.
}
\examples{
Y <- gen_Nmix_closed(5,5,250,0.5)
plot_red_like_closed_lambda(Y=reduction(Y$nit,10), startLambda = 150, endLambda = 350, stepsize=10, pdet = 0.5, red = 10, K = reduction(400,10))
}
