% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_reduced_count_models.R
\name{plot_red_like_closed_pdet}
\alias{plot_red_like_closed_pdet}
\title{Plot likelihood given a lambda and range for pdet.}
\usage{
plot_red_like_closed_pdet(nit, startPdet, endPdet, stepsize, lambda, red,
  K)
}
\arguments{
\item{nit}{R by T matrix of reduced counts with R sites/rows and T sampling occassions/columns.}

\item{startPdet}{Starting value for pdet.}

\item{endPdet}{Ending value for pdet.}

\item{stepsize}{Spacing between values of pdet}

\item{lambda}{Initial abundance parameter.}

\item{red}{reduction factor.}

\item{K}{Upper bound on summations (reduced counts upper bound).}
}
\description{
Plot likelihood given a lambda and range for pdet.
}
\examples{
Y <- gen_Nmix_closed(5,5,250,0.5)
plot_red_like_closed_pdet(Y=reduction(Y$nit,10), startPdet = 0.1, endPdet = 1.0, stepsize=0.1, lambda = 250, red = 10, K = reduction(400,10))
}
