% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{rbind_pops}
\alias{rbind_pops}
\title{rbind_pops}
\usage{
rbind_pops(pop1, pop2)
}
\arguments{
\item{pop1}{first population list, must have named matrices Ni and nit as list elements.}

\item{pop2}{second population list, must have named matrices Ni and nit with same number of columns (sampling occasions) as pop1.}
}
\value{
Single population list with sites from pop1 and pop2.
}
\description{
rbind two population lists (as generated from either gen_Nmix_closed or gen_Nmix_open)
}
