% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_reduced_count_models.R
\name{reduction}
\alias{reduction}
\title{Reduce an integer value using a reduction function \eqn{R(x;r)}. Currently only implements round2(), which is standard rounding (\{x\}.5 rounds to \{x+1\}.0).}
\usage{
reduction(x, red, FUN = round2)
}
\arguments{
\item{x}{The integer which is to be reduced.}

\item{red}{The factor r by which to reduce the input x.}

\item{FUN}{The reduction function (default is round2(), some alternatives are ceiling() and floor())}
}
\value{
An integer value which is the reduction of integer x by reduction factor red using function FUN.
}
\description{
Reduce an integer value using a reduction function \eqn{R(x;r)}. Currently only implements round2(), which is standard rounding (\{x\}.5 rounds to \{x+1\}.0).
}
\examples{
x <- 104
xr1 <- reduction(x, 10)
xr2 <- reduction(x, 10, ceiling)
}
