% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions.R
\name{round2}
\alias{round2}
\title{round2}
\usage{
round2(x, n = 0)
}
\arguments{
\item{x}{number to be rounded}

\item{n}{decimal place to round to (n=0 rounds to nearest whole number, n=1 rounds to the nearest tenth, n=-1 rounds to the nearest tens digit, etc).}
}
\description{
rounding in R rounds to nearest even number (eg 0.5 rounds to 0), this function does "normal" rounding (eg 0.5 rounds to 1)
}
