% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_reduced_count_models.R
\name{START_PARALLEL}
\alias{START_PARALLEL}
\title{Used to initialize parallel computing (useful for likelihood calculations which can be computationally intensive).}
\usage{
START_PARALLEL(num_cores)
}
\arguments{
\item{num_cores}{Number of cores to use for parallel processing.}
}
\description{
Used to initialize parallel computing (useful for likelihood calculations which can be computationally intensive).
}
