% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_reduced_count_models.R
\name{tp_jk_V}
\alias{tp_jk_V}
\title{Internal function, calculates transition probabilities from pop size j to pop size k in the open population likelihood.
can this be run as a parallel process using multiple cores?}
\usage{
tp_jk_V(j_vec, k_vec, omeg, gamm, red)
}
\description{
Internal function, calculates transition probabilities from pop size j to pop size k in the open population likelihood.
can this be run as a parallel process using multiple cores?
}
