% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_reduced_count_models.R
\name{tp_MAT}
\alias{tp_MAT}
\title{Internal function, calculates transition probability matrix (transition from row pop to column pop)}
\usage{
tp_MAT(M, omeg, gamm, red)
}
\description{
Internal function, calculates transition probability matrix (transition from row pop to column pop)
}
