% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_reduced_count_models.R
\name{tp_MAT_APA}
\alias{tp_MAT_APA}
\title{Internal function, calculates transition probability matrix (transition from row pop to column pop)}
\usage{
tp_MAT_APA(M, omeg, B_o, omet, B_ot, gamm, B_g, gamt, B_gt, red,
  PARALLELIZE = FALSE, precBits = 128)
}
\description{
Internal function, calculates transition probability matrix (transition from row pop to column pop)
}
