% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{drbinom}
\alias{drbinom}
\title{drbinom}
\usage{
drbinom(x, size, prob, red, log = FALSE)
}
\arguments{
\item{x}{Reduced count quantile (alternatively input reduction(x,red) if x is a full count quantile).}

\item{size}{Number of trials (full count size).}

\item{prob}{Probability of success for each trial.}

\item{red}{The factor r by which x has been reduced.}
}
\value{
The probability of observing quantile \eqn{x}.
}
\description{
Reduced binomial probability distribution function \eqn{rBinomial(x;N,p,R(x;r))}, takes reduced quantiles rather than full quantiles.
}
\examples{
Y <- drbinom(0:20, 20, 0.3, 2)
plot(Y, xlab="Y~rBinom(N=20,p=0.3,r=2)", ylab="P[Y=y]")
}
