% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_reduced_count_models.R
\name{drbinom}
\alias{drbinom}
\title{Reduced binomial probability distribution function \eqn{rBinomial(x;N,p,R(x;r))}}
\usage{
drbinom(x, size, prob, red, log = FALSE)
}
\arguments{
\item{x}{Full count quantile (alternatively input r*x if x is a reduced count quantile).}

\item{size}{Number of trials.}

\item{prob}{Probability of success for each trial.}

\item{red}{The factor r by which to reduce the input x.}
}
\value{
The probability of observing quantile \eqn{R(x;r)}
}
\description{
Reduced binomial probability distribution function \eqn{rBinomial(x;N,p,R(x;r))}
}
\examples{
Y <- drbinom(0:200, 200, 0.3, 10)[seq(1,201,10)]
plot(Y, xlab="Y=R(X;10), X~Binomial(N=200,p=0.3)", ylab="P[Y=y]")
}
