% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{drbinomAPA}
\alias{drbinomAPA}
\title{drbinomAPA}
\usage{
drbinomAPA(x, size, prob, red, precBits = 128, log = FALSE)
}
\arguments{
\item{x}{Reduced count quantile (alternatively input reduction(x,r) if x is a full count quantile).}

\item{size}{Number of trials.}

\item{prob}{Probability of success for each trial.}

\item{red}{The factor r by which x has been reduced.}

\item{precBits}{Number of bits of precision for arbitrary precision arithmetic.}
}
\value{
The probability of observing quantile \eqn{x}.
}
\description{
Arbitrary precision reduced binomial probability distribution function \eqn{rBinomial(x;N,p,R(x;r))}, takes reduced quantiles rather than full quantiles.
}
\examples{
Y <- drbinomAPA(0:3, 20, 0.3, 10, precBits=64)
Y
}
