% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_reduced_count_models.R
\name{drpois}
\alias{drpois}
\title{Reduced poisson probability distribution function rPoisson(x;lambda,r)}
\usage{
drpois(x, lambda, red, K = 0)
}
\arguments{
\item{x}{Full count quantile (alternatively input r*x if x is a reduced count quantile).}

\item{lambda}{Mean of the full count poisson distribution.}

\item{red}{The factor r by which to reduce the input x.}

\item{K}{Upper bound after which to truncate Poisson (used in calculating alpha to normalize rPoisson).
Default K=0 to have upper bound calculated automatically (not optimized).}
}
\value{
The probability of observing quantile $R(x;r)$
}
\description{
Reduced poisson probability distribution function rPoisson(x;lambda,r)
}
\examples{
Y <- drpois(seq(1,200,10), 55, 10)
plot(Y, xlab="Y=R(X;10)", main="X~Poisson(lambda=55)", ylab="P[Y=y]")
}
