% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_reduced_count_models.R
\name{fit_red_Nmix_closed}
\alias{fit_red_Nmix_closed}
\title{Find maximum likelihood estimates for model parameters log(lambda) and logit(pdet). Uses optim.}
\usage{
fit_red_Nmix_closed(nit, red, K, starts = c(1, 0), VERBOSE = FALSE,
  method = "BFGS", ...)
}
\arguments{
\item{nit}{R by T matrix of full counts with R sites/rows and T sampling occassions/columns.}

\item{red}{reduction factor.}

\item{K}{Upper bound on summations (full count value, eg if K=300 for full counts, K=reduction(300,red)+1 for reduced counts).}

\item{starts}{Vector of starting values for optimize. Has two elements, log(lambda) and logit(pdet).}

\item{VERBOSE}{If true, prints the log likelihood to console at each optim iteration.}

\item{...}{Additional input for optim.}
}
\description{
Find maximum likelihood estimates for model parameters log(lambda) and logit(pdet). Uses optim.
}
\examples{
Y <- gen_Nmix_closed(1,5,250,0.5)
out <- fit_red_Nmix_closed(Y$nit, red=10, K=300, starts = c(log(250),boot::logit(0.5)))
}
