% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_reduced_count_models.R
\name{gen_Nmix_open}
\alias{gen_Nmix_open}
\title{Generate a population/observation pair with the structure of an open N-mixture model}
\usage{
gen_Nmix_open(num_sites, num_times, lambda, pdet, omega, gamma)
}
\arguments{
\item{num_sites}{The number of observation sites.}

\item{num_times}{The number of sampling occasions.}

\item{lambda}{The population rate parameter (\eqn{N_i \sim Poisson(\lambda)}).}

\item{pdet}{The probability of detection \eqn{p} (\eqn{n_{it} \sim Binomial(N_i,p)}).}

\item{omega}{The probability of survival (\eqn{S_{it} \sim Binomial(N_{it}, \omega)}). Either a single number, or a vector of length t-1 (omega for each sampling occassion after the first).}

\item{gamma}{The recruitment rate parameter (\eqn{G_{it} \sim Poisson(\gamma)}). Either a single number, or a vector of length t-1 (gamma for each sampling occassion after the first).}
}
\value{
A list object with two named matrices. Nit contains the total population per site
        (each row represents a site, each column a sampling occasion). nit contains the observed
        counts (rows=sites, columns=sampling occasions).
}
\description{
Generate a population/observation pair with the structure of an open N-mixture model
}
\examples{
pop <- gen_Nmix_open(num_sites = 5,
                     num_times = 10,
                     lambda = 50,
                     pdet = 0.4,
                     omega = 0.8,
                     gamma = 4)
pop
}
