% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_reduced_count_models.R
\name{max_r}
\alias{max_r}
\title{Recommend a reduction value \eqn{r} based on maximum threshold on increased variance.}
\usage{
max_r(nit, threshold = 1.1)
}
\arguments{
\item{nit}{Full counts.}

\item{threshold}{Desired maximum ratio of \eqn{v2/v1}, where \eqn{v2} is the variance of the full counts,
and \eqn{v1} is the variance of the reduced counts, scaled by \eqn{r^2}.}
}
\value{
Recommended maximum reduction factor r to use when fitting reduced count model with given sample nit.
}
\description{
Recommend a reduction value \eqn{r} based on maximum threshold on increased variance.
}
\examples{
Y <- redNMix::gen_Nmix_closed(num_sites = 5,num_times = 5,lambda = 250,pdet = 0.5)
max_r(nit = Y$nit)

Y <- redNMix::gen_Nmix_closed(num_sites = 5,num_times = 5,lambda = 2500,pdet = 0.15)
z <- lapply(X = 1:200, FUN = redNMix::reduction, x=Y$nit)
z2 <- lapply(X = z, FUN = function(x){var(as.numeric(x))})
plot(sqrt(unlist(z2))*1:200,
     xlab="reduction factor r",
     ylab="standard deviation of scaled reduced counts")
abline(h = sqrt(1.1*var(as.numeric(unlist(z[[1]])))))
title("sd(nit_r) for r 1:200")
}
