% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions.R
\name{mcd}
\alias{mcd}
\title{mcd}
\usage{
mcd(nit, howmany = 10)
}
\arguments{
\item{nit}{Full counts}

\item{howmany}{Maximum number of common divisors to return.}
}
\description{
Find the Most Common Divisor of a data set
}
\examples{
mcd(c(10,20,30,35,12,18), howmany = 5)

}
