% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_2d_red_like_closed}
\alias{plot_2d_red_like_closed}
\title{plot_2d_red_like_closed}
\usage{
plot_2d_red_like_closed(nit, startPdet, endPdet, stepsizePdet, startLambda,
  endLambda, stepsizeLambda, red, K)
}
\arguments{
\item{nit}{R by T matrix of reduced counts with R sites/rows and T sampling occassions/columns.}

\item{startPdet}{Starting value for pdet.}

\item{endPdet}{Ending value for pdet.}

\item{stepsizePdet}{Spacing between values of pdet}

\item{startLambda}{Starting value for lambda.}

\item{endLambda}{Ending value for lambda.}

\item{stepsizeLambda}{Spacing between values of lambda}

\item{red}{Reduction factor r.}

\item{K}{Upper bound on summations (reduced counts upper bound).}
}
\value{
Returns a matrix of likelihoods where rows represent pdet, and columns represent lambda.
}
\description{
Plot 2D closed likelihood given a range for pdet and for lambda.
}
\examples{
set.seed(12345)
Y <- gen_Nmix_closed(15,15,150,0.5)
L <- plot_2d_red_like_closed(reduction(Y$nit,10), 0.05, 1.0, 0.05, 10, 250, 10, 10, reduction(400,10))
}
