% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red_Like_closed.R
\name{red_Like_closed}
\alias{red_Like_closed}
\title{red_Like_closed}
\usage{
red_Like_closed(par, nit, l_s_c, p_s_c, K, red, FUN = round,
  VERBOSE = FALSE, PARALLELIZE = FALSE, APA = FALSE, precBits = 64)
}
\arguments{
\item{par}{Vector with two elements, log(lambda) and logis(pdet). If l_s_c is not NULL, need length(par) = length(l_s_c) + 2, for the B0...BK coefficients of lambda (B0 is the constant term coefficient).}

\item{nit}{R by T matrix of reduced counts with R sites/rows and T sampling occassions/columns.}

\item{l_s_c}{list of lambda site covariates (list of vectors of length R (number of sites), one vector per covariate)}

\item{p_s_c}{list of pdet site covariates (list of vectors of length R (number of sites), one vector per covariate)}

\item{K}{Upper bound on summations (input the reduced count upper bound). An R by T matrix, eg: K=matrix(50, nrow=R, ncol=T).}

\item{red}{reduction factor matrix (R by T, with R sites/rows and T sampling occassions/columns)}

\item{VERBOSE}{If TRUE, prints the calculated log likelihood to console.}

\item{PARALLELIZE}{If TRUE, calculations will be done in parallel. Will use as many cores as have been made available (initialize with START_PARALLEL(num_cores)).}

\item{APA}{Default is FALSE. If TRUE, will use arbitrary precision arithmetic in calculating the likelihood. Note that APA will be slower, however it is required for site population sizes larger than about 300. If APA = TRUE, then precBits specifies the number of bits of precision to use in the calculations.}

\item{precBits}{If APA=TRUE, then precBits specifies the number of bits of precision for arbitrary precision arithmetic.}
}
\description{
Used to calculate the negative of the log likelihood for closed population models.
}
