% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_reduced_count_models.R
\name{red_Like_closed}
\alias{red_Like_closed}
\title{Internal function, should not be called directly. Used to calculate the negative of the likelihood.}
\usage{
red_Like_closed(par, nit, K, red, FUN = round)
}
\arguments{
\item{par}{Vector with two elements, logis(pdet) and lambda.}

\item{nit}{R by T matrix of full counts with R sites/rows and T sampling occassions/columns.}

\item{K}{Upper bound on summations.}

\item{red}{reduction factor}
}
\description{
Internal function, should not be called directly. Used to calculate the negative of the likelihood.
}
