% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_reduced_count_models.R
\name{red_Like_closed}
\alias{red_Like_closed}
\title{Internal function. Used to calculate the negative of the log likelihood.}
\usage{
red_Like_closed(par, nit, K, red, FUN = round, VERBOSE = FALSE,
  PARALLELIZE = FALSE)
}
\arguments{
\item{par}{Vector with two elements, log(lambda) and logis(pdet).}

\item{nit}{data.frame with R*T rows, and 3 columns: "site", "time", and "count". Deprecated: R by T matrix of reduced counts with R sites/rows and T sampling occassions/columns.}

\item{K}{Upper bound on summations (input reduced count upper bound).}

\item{red}{reduction factor}

\item{VERBOSE}{If true, prints the calculated log likelihood to console.}

\item{PARALLELIZE}{If true, will use as many threads as have been made available (initialize with START_PARALLEL(num_cores)).}
}
\description{
Internal function. Used to calculate the negative of the log likelihood.
}
