% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_reduced_count_models.R
\name{red_Like_open}
\alias{red_Like_open}
\title{Internal function. Used to calculate the negative of the log likelihood.}
\usage{
red_Like_open(par, nit, K, red, FUN = round)
}
\arguments{
\item{par}{Vector with four elements, logis(pdet), log(lambda), logis(omega), and log(gamma).}

\item{nit}{R by T matrix of full counts with R sites/rows and T sampling occassions/columns.}

\item{K}{Upper bound on summations.}

\item{red}{reduction factor}
}
\description{
Internal function. Used to calculate the negative of the log likelihood.
}
