% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_reduced_count_models.R
\name{red_Like_open}
\alias{red_Like_open}
\title{Internal function. Used to calculate the negative of the log likelihood.}
\usage{
red_Like_open(par, nit, K, red, FUN = round, VERBOSE = FALSE)
}
\arguments{
\item{par}{Vector with four elements, log(lambda), log(gamma), logis(omega), and logis(pdet).}

\item{nit}{R by T matrix of reduced counts with R sites/rows and T sampling occassions/columns.}

\item{K}{Upper bound on summations (reduced counts upper bound).}

\item{red}{Reduction factor}

\item{VERBOSE}{If true, prints the log likelihood to console.}
}
\description{
Internal function. Used to calculate the negative of the log likelihood.
}
\details{
Note that this function is adapted from the negative log likelihood function from the unmarked package, and uses the recursive method of computation described in Web Appendix A of Dail and Madsen 2011: Models for Estimating Abundance from Repeated Counts of an Open Metapopulation, published in Biometrics volume 67, issue 2.
}
